fix: allow decimal refrigerant quantities and fix BoQ legend asterisk - Change refrigerant_quantity_kg from PositiveIntegerField to DecimalField(max_digits=8, decimal_places=2) on CoolingSystemChiller and CoolingSystemAirConditioner; add migration 0029                        - Reword BoQ/design drawings legend so asterisk stays on same line

+++ pages/models/building_operation/air_conditioning.py
@@ -53,7 +53,8 @@ class CoolingSystemAirConditioner(models.Model):
         choices=RefrigerantType.choices,
         verbose_name=_("Type of Refrigerants"),
     )
-    refrigerant_quantity_kg = models.PositiveIntegerField(
+    refrigerant_quantity_kg = models.DecimalField(
+        max_digits=8, decimal_places=2,
         verbose_name=_("Refrigerant Quantity (Kg)")
     )
     total_cooling_load_rt = models.PositiveIntegerField(
